% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fnGetAMat}
\alias{fnGetAMat}
\title{Matrix with each element a_{ij} = I(\\sum_{k=1}^K \\delta_{ikj} \\ne 0)}
\usage{
fnGetAMat(Delta.mat, subject)
}
\description{
Matrix with each element a_{ij} = I(\\sum_{k=1}^K \\delta_{ikj} \\ne 0)
}
