% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fnGetBMat}
\alias{fnGetBMat}
\title{Matrix with each element b_{kj} = I(\\sum_{i=1}^n \\delta_{ikj} \\ne 0)}
\usage{
fnGetBMat(Delta.mat, condition)
}
\description{
Matrix with each element b_{kj} = I(\\sum_{i=1}^n \\delta_{ikj} \\ne 0)
}
