% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fnGetDeltaSumMat}
\alias{fnGetDeltaSumMat}
\title{Matrix with each element is \\Delta_{kj} = \\sum_{k^\\prime \\ne k} b_{kj}}
\usage{
fnGetDeltaSumMat(B.mat)
}
\description{
Matrix with each element is \\Delta_{kj} = \\sum_{k^\\prime \\ne k} b_{kj}
}
