% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fnGetMuMat}
\alias{fnGetMuMat}
\title{Calculate Mu.mat from the other parameters}
\usage{
fnGetMuMat(r.vec, Alpha.mat, Theta.mat, Delta.mat, s.vec, i.vec, k.vec,
  X.mat = NA, Beta.mat = NA)
}
\description{
Calculate Mu.mat from the other parameters
}
