% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{fnSampPsi}
\alias{fnSampPsi}
\title{Sample Psi}
\usage{
fnSampPsi(d.vec, a.psi)
}
\arguments{
\item{d.vec}{vector of length L of cluster sizes}

\item{a.psi}{from the prior, vector of length L}
}
\description{
Sample Psi
}
