% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpm.R
\name{fnSampVVec}
\alias{fnSampVVec}
\title{Sample stick-breaking weights}
\usage{
fnSampVVec(rho, d.vec, d.cumsum.vec)
}
\arguments{
\item{rho}{DP concentration parameter}
}
\description{
Sample stick-breaking weights
}
