% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fnGetMuMat}
\alias{fnGetMuMat}
\title{Calculate Mu.mat from the other parameters}
\usage{
fnGetMuMat(r.vec, Alpha.mat, Beta.mat, Delta.mat, s.vec)
}
\description{
Calculate Mu.mat from the other parameters
}
